Flask boilerplate code
Includes a class for PyPew, and a create_app() function that creates a
Flask app object, and a main() for running the app. Server configuration
is read from a .env file, and a simple unit test and a pair of templates
are provided, as well as a minimal requirements.txt file.

In the future it may be desirable to define the view functions at the
top level, not as inner functions of create_app().

Test plan: Run the unit test.

References:

  [1] https://stackoverflow.com/a/45017691