.docx generation, and frontend for .pdf
Buttons added to service view to download .docx and .pdf. You can
generate a minimalist .docx that has a heading with the name of the
service.

Not quite Issue #1 since this generates a new .docx file rather than
using a template.

.pdf generation doesn't work yet, and from looking at docx2pdf it seems
that it will require separate testing on Windows, Linux and Mac.