Add a .xlsx file for some test cases
This commit adds a demo .xlsx file containing some texts (readings,
propers, etc.) to the repository. The file reflects the format in which
the user would supply a full .xlsx file with all the texts. One row for
each test case (combinations of Gradual, Alleluia and Tract).

A copyright statement has been added to README.md.