Make the pew sheet and feast views printable
With this change the pew sheet view (/pewSheet/?title=...) and the feast
detail view (/feast/Advent I) are in a friendlier form for printing
(with no sidebar, buttons, etc.). There is also a button on each view to
invoke the print dialog (which on Mac allows you to save the page to a
PDF).

This is intended to be a temporary (and lightweight) alternative to
creating a .docx that can be properly edited.